Change class hierarchy, add movement classes.
In this commit we added a change in the class hierarchy,
as before there was a PlayerDroid class, now this is
just a regular Droid with a toggle on. We added four new
classes, DroidMovement, an abstract class in which a
droid variable is declared and a constructor, as well
as an abstract method, Move(), PlayerDroidMovement,
an extension of DroidMovement with inherited Move(),
AIDroidMovement, another extension with an inherited
Move(), and Global,  which isn't finished yet but will
hold all Global game info. I also deleted the PlayerDroid
file, and merged most of its functonality into Droid and
the DroidMovement classes.

TODO
- Finish player movement and AI movement.
- Finish Global class.
- Add projectiles.
- Eventually add hacking mechanic.
- Add ability to win the game.