Added basic saving and loading.
In this commit we added a basic saving and loading feature. This
isn't quite usable yet, but it will allow for the player to save
their current level and stats, as well as all of the enemies in
the map. There is a GameInformation class, in which we will add
all of the information to save, a GameInformationHolder class,
which will be applied to an object and hold the GameInformation
class. The SaveAndLoad class is full of static methods to help
save, serialize and load the GameInformation class.

TODO
- Allow for the saving and loading to be usable.
- Main menu.
- Enemy AI.
- Destroying enemies.
- Hacking.
- Winning the game.
- Player models.
- Sound and music.